integration_package/backend/requirements.txt
# CHE·NU Backend Requirements
# Python 3.11+

# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Redis & Caching
redis>=5.0.1
aioredis>=2.0.1

# Pydantic & Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.1

# AI/LLM
anthropic>=0.18.0
openai>=1.10.0

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0.1
structlog>=24.1.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
black>=24.1.0
isort>=5.13.2
mypy>=1.8.0
ruff>=0.1.14
